Add ranking formulae
