--- TowerForge/TowerForge/Commons/Enums/RankingEnums.swift
@@ -1,22 +1,0 @@
-//
-//  RankingEnums.swift
-//  TowerForge
-//
-//  Created by Rubesh on 16/4/24.
-//
-
-import Foundation
-
-class RankingEnums {
-
-    enum Rank: String, CaseIterable {
-        case PRIVATE
-        case CORPORAL
-        case SERGEANT
-        case LIEUTENANT
-        case CAPTAIN
-        case MAJOR
-        case COLONEL
-        case GENERAL
-    }
-}
